Correct fold behavior in angelic evaluator
The implementation was checking whether a fold's destination object was self to
determine whether a final equality assertion was necessary, when in reality,
any singleton child could be the destination object. Because the type-checker
ruled out any ill-formed attribute references or assignments (w.r.t. loops and
indices), we can assume that if the destination object is not the iterated
child sequence, it must be either self or a singleton object, thus requiring a
post-fold assertion of equality with the final accumulator value.